+++ CHANGELOG.md
@@ -15,3 +15,51 @@
   - Get-WarrantyOptionsBySerialAndCountryCode
   - Get-WarrantyOptionsByMachineType
   - Get-WarrantyOptionsByMachineTypeAndCountryCode
+
+## [0.0.2] - 2019-07-28
+### Added
+- Product Catalog Endpoints
+  - Get-Catalog
+  - Get-Catalog-Raw
+  - Get-Catalog-All
+  - Get-Catalog-Children
+
+- Product Information Endpoint(s)
+  - Get-Product
+
+## [0.0.3] - 2019-08-03
+### Added
+- Content
+  - Get-Content
+
+- Support
+  - Get-Support-ServiceRequests
+  - Get-Support-ServiceOrders
+
+- Part
+  - Get-Part
+  - Get-PartByIds
+
+## [0.0.4] - 2019-08-04
+### Added
+- Accessory
+  - Get-Accessory
+  - Get-AccessoryByIds
+  - Get-AccessoryForProduct
+  - Get-AccessoriesFromSerial
+
+- Provider
+  - Get-Provider-ServiceProviders
+
+- Contact
+  - Get-Contacts-SupportContacts
+
+## [0.0.5] - 2019-08-11
+### Added
+- Case
+  - Get-Case
+  - Get-CasesByIds
+  - New-Case
+
+- Profile
+  - Get-Profile